requirements.txt
garminconnect>=0.1.30
pyyaml>=6.0